Save Flux_scalars, the average value of each model spectrum, required for absolute flux of models and accurate solid angles.
Note that this is backwards-INCOMPATIBLE.  We have to pass an extra positional argument to many of the classes and methods that are not expecting one.  So trying to use this code with previous versions will fail.  This change is not desirable but the old code was misleading in the solid angle support, so it can be considered a bug fix.